+++ dflib-json/src/main/java/org/dflib/json/JsonLoader.java
@@ -4,18 +4,11 @@ import com.jayway.jsonpath.Configuration;
 import com.jayway.jsonpath.DocumentContext;
 import com.jayway.jsonpath.JsonPath;
 import com.jayway.jsonpath.Option;
-import org.dflib.BoolValueMapper;
 import org.dflib.DataFrame;
-import org.dflib.DoubleValueMapper;
 import org.dflib.Extractor;
-import org.dflib.IntValueMapper;
-import org.dflib.LongValueMapper;
 import org.dflib.ValueMapper;
 
 import java.io.Reader;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
 import java.util.HashMap;
 import java.util.HashSet;
@@ -30,7 +23,10 @@ import java.util.Set;
 public class JsonLoader {
 
     private final Set<Option> options;
-    private final Map<String, Extractor<Map<String, Object>, ?>> extractorPresets;
+
+    // unlike CSV and other columnar data sources, JSON does not have column positions, so custom
+    // extractors can be resolved immediately and stored in the map by column name
+    private final Map<String, Extractor<Map<String, Object>, ?>> extractors;
 
     private String pathExpression;
 
@@ -39,7 +35,7 @@ public class JsonLoader {
         this.options = new HashSet<>();
         options.add(Option.ALWAYS_RETURN_LIST);
 
-        this.extractorPresets = new HashMap<>();
+        this.extractors = new HashMap<>();
     }
 
     /**
@@ -55,6 +51,10 @@ public class JsonLoader {
         return this;
     }
 
+    /**
+     * If set, prevents skipping rows that are missing properties defined in the path expression. Instead
+     * of skipping those rows, the loader would fill those positions with null.
+     */
     public JsonLoader nullsForMissingLeafs() {
         this.options.add(Option.DEFAULT_PATH_LEAF_TO_NULL);
         return this;
@@ -64,7 +64,27 @@ public class JsonLoader {
      * @since 1.0.0-RC1
      */
     public JsonLoader col(String column, ValueMapper<Object, ?> mapper) {
-        extractorPresets.put(column, customExtractor(column, mapper));
+        extractors.put(column, ColConfigurator.objectCol(column, mapper, false));
+        return this;
+    }
+
+    /**
+     * Configures a column to be loaded with value compaction. Should be used to save memory for low-cardinality columns.
+     *
+     * @since 1.0.0-RC1
+     */
+    public JsonLoader compactCol(String column) {
+        extractors.put(column, ColConfigurator.objectCol(column, true));
+        return this;
+    }
+
+    /**
+     * Configures a column to be loaded with value compaction. Should be used to save memory for low-cardinality columns.
+     *
+     * @since 1.0.0-RC1
+     */
+    public JsonLoader compactCol(String column, ValueMapper<Object, ?> mapper) {
+        extractors.put(column, ColConfigurator.objectCol(column, mapper, true));
         return this;
     }
 
@@ -72,7 +92,7 @@ public class JsonLoader {
      * @since 0.16
      */
     public JsonLoader boolCol(String column) {
-        extractorPresets.put(column, boolExtractor(column));
+        extractors.put(column, ColConfigurator.boolCol(column));
         return this;
     }
 
@@ -80,7 +100,7 @@ public class JsonLoader {
      * @since 0.16
      */
     public JsonLoader intCol(String column) {
-        extractorPresets.put(column, intExtractor(column));
+        extractors.put(column, ColConfigurator.intCol(column));
         return this;
     }
 
@@ -88,7 +108,7 @@ public class JsonLoader {
      * @since 0.16
      */
     public JsonLoader intCol(String column, int forNull) {
-        extractorPresets.put(column, intExtractor(column, forNull));
+        extractors.put(column, ColConfigurator.intCol(column, forNull));
         return this;
     }
 
@@ -96,7 +116,7 @@ public class JsonLoader {
      * @since 0.16
      */
     public JsonLoader longColumn(String column) {
-        extractorPresets.put(column, longExtractor(column));
+        extractors.put(column, ColConfigurator.longCol(column));
         return this;
     }
 
@@ -104,7 +124,7 @@ public class JsonLoader {
      * @since 0.16
      */
     public JsonLoader longColumn(String column, long forNull) {
-        extractorPresets.put(column, longExtractor(column, forNull));
+        extractors.put(column, ColConfigurator.longCol(column, forNull));
         return this;
     }
 
@@ -112,7 +132,7 @@ public class JsonLoader {
      * @since 0.16
      */
     public JsonLoader doubleCol(String column) {
-        extractorPresets.put(column, doubleExtractor(column));
+        extractors.put(column, ColConfigurator.doubleCol(column));
         return this;
     }
 
@@ -120,7 +140,7 @@ public class JsonLoader {
      * @since 0.16
      */
     public JsonLoader doubleCol(String column, double forNull) {
-        extractorPresets.put(column, doubleExtractor(column, forNull));
+        extractors.put(column, ColConfigurator.doubleCol(column, forNull));
         return this;
     }
 
@@ -128,7 +148,7 @@ public class JsonLoader {
      * @since 0.16
      */
     public JsonLoader dateCol(String column) {
-        extractorPresets.put(column, dateExtractor(column));
+        extractors.put(column, ColConfigurator.dateCol(column));
         return this;
     }
 
@@ -136,7 +156,7 @@ public class JsonLoader {
      * @since 0.16
      */
     public JsonLoader dateCol(String column, DateTimeFormatter formatter) {
-        extractorPresets.put(column, dateExtractor(column, formatter));
+        extractors.put(column, ColConfigurator.dateCol(column, formatter));
         return this;
     }
 
@@ -144,7 +164,7 @@ public class JsonLoader {
      * @since 0.16
      */
     public JsonLoader timeCol(String column) {
-        extractorPresets.put(column, timeExtractor(column));
+        extractors.put(column, ColConfigurator.timeCol(column));
         return this;
     }
 
@@ -152,7 +172,7 @@ public class JsonLoader {
      * @since 0.16
      */
     public JsonLoader timeCol(String column, DateTimeFormatter formatter) {
-        extractorPresets.put(column, timeExtractor(column, formatter));
+        extractors.put(column, ColConfigurator.timeCol(column, formatter));
         return this;
     }
 
@@ -160,7 +180,7 @@ public class JsonLoader {
      * @since 0.16
      */
     public JsonLoader dateTimeCol(String column) {
-        extractorPresets.put(column, dateTimeExtractor(column));
+        extractors.put(column, ColConfigurator.dateTimeCol(column));
         return this;
     }
 
@@ -168,10 +188,9 @@ public class JsonLoader {
      * @since 0.16
      */
     public JsonLoader dateTimeCol(String column, DateTimeFormatter formatter) {
-        extractorPresets.put(column, dateTimeExtractor(column, formatter));
+        extractors.put(column, ColConfigurator.dateTimeCol(column, formatter));
         return this;
     }
-
 
     public DataFrame load(String json) {
         DocumentContext context = JsonPath.parse(json, buildJSONPathConfiguration());
@@ -183,95 +202,12 @@ public class JsonLoader {
         return load(context);
     }
 
+    protected DataFrame load(DocumentContext context) {
+        List<?> parsed = context.read(pathExpression);
+        return new JsonLoaderWorker(extractors).load(parsed);
+    }
+
     protected Configuration buildJSONPathConfiguration() {
         return Configuration.builder().options(options).build();
-    }
-
-    protected DataFrame load(DocumentContext context) {
-        List<Map<String, Object>> parsed = context.read(pathExpression);
-        return new JsonLoaderWorker(extractorPresets).load(parsed);
-    }
-
-    static Extractor<Map<String, Object>, Object> defaultExtractor(String name) {
-        return Extractor.$col(m -> getObject(m, name));
-    }
-
-    static Extractor<Map<String, Object>, Object> customExtractor(String name, ValueMapper<Object, ?> mapper) {
-        return Extractor.$col(m -> mapper.map(getObject(m, name)));
-    }
-
-    static Extractor<Map<String, Object>, Boolean> boolExtractor(String name) {
-        return Extractor.$bool(m -> BoolValueMapper.fromObject().map(getObject(m, name)));
-    }
-
-    static Extractor<Map<String, Object>, Integer> intExtractor(String name) {
-        return Extractor.$int(m -> IntValueMapper.fromObject().map(getObject(m, name)));
-    }
-
-    static Extractor<Map<String, Object>, Integer> intExtractor(String name, int fillNulls) {
-        IntValueMapper<Object> mapper = IntValueMapper.fromObject(fillNulls);
-        return Extractor.$int(m -> mapper.map(getObject(m, name)));
-    }
-
-    static Extractor<Map<String, Object>, Long> longExtractor(String name) {
-        return Extractor.$long(m -> LongValueMapper.fromObject().map(getObject(m, name)));
-    }
-
-    static Extractor<Map<String, Object>, Long> longExtractor(String name, long fillNulls) {
-        LongValueMapper<Object> mapper = LongValueMapper.fromObject(fillNulls);
-        return Extractor.$long(m -> mapper.map(getObject(m, name)));
-    }
-
-    static Extractor<Map<String, Object>, Double> doubleExtractor(String name) {
-        return Extractor.$double(m -> DoubleValueMapper.fromObject().map(getObject(m, name)));
-    }
-
-    static Extractor<Map<String, Object>, Double> doubleExtractor(String name, double fillNulls) {
-        DoubleValueMapper<Object> mapper = DoubleValueMapper.fromObject(fillNulls);
-        return Extractor.$double(m -> mapper.map(getObject(m, name)));
-    }
-
-    static Extractor<Map<String, Object>, LocalDate> dateExtractor(String name) {
-        return Extractor.$col(m -> ValueMapper.stringToDate().map(getString(m, name)));
-    }
-
-    static Extractor<Map<String, Object>, LocalDate> dateExtractor(String name, DateTimeFormatter formatter) {
-        ValueMapper<String, LocalDate> mapper = ValueMapper.stringToDate(formatter);
-        return Extractor.$col(m -> mapper.map(getString(m, name)));
-    }
-
-    static Extractor<Map<String, Object>, LocalTime> timeExtractor(String name) {
-        return Extractor.$col(m -> ValueMapper.stringToTime().map(getString(m, name)));
-    }
-
-    static Extractor<Map<String, Object>, LocalTime> timeExtractor(String name, DateTimeFormatter formatter) {
-        ValueMapper<String, LocalTime> mapper = ValueMapper.stringToTime(formatter);
-        return Extractor.$col(m -> mapper.map(getString(m, name)));
-    }
-
-    static Extractor<Map<String, Object>, LocalDateTime> dateTimeExtractor(String name) {
-        return Extractor.$col(m -> ValueMapper.stringToDateTime().map(getString(m, name)));
-    }
-
-    static Extractor<Map<String, Object>, LocalDateTime> dateTimeExtractor(String name, DateTimeFormatter formatter) {
-        ValueMapper<String, LocalDateTime> mapper = ValueMapper.stringToDateTime(formatter);
-        return Extractor.$col(m -> mapper.map(getString(m, name)));
-    }
-
-    private static Object getObject(Map<String, Object> map, String key) {
-        if (map == null) {
-            return null;
-        }
-
-        return map.get(key);
-    }
-
-    private static String getString(Map<String, Object> map, String key) {
-        if (map == null) {
-            return null;
-        }
-
-        Object val = map.get(key);
-        return val != null ? val.toString() : null;
     }
 }