+++ dflib-json/src/main/java/org/dflib/json/JsonLoaderWorker.java
@@ -16,23 +16,23 @@ class JsonLoaderWorker {
 
     private static final String DEFAULT_SCALAR_COLUMN = "_val";
 
-    private final Map<String, Extractor<Map<String, Object>, ?>> extractorPresets;
+    private final Map<String, Extractor<Map<String, Object>, ?>> extractors;
     private final LinkedHashMap<String, SeriesAppender<Map<String, Object>, ?>> appenders;
 
-    JsonLoaderWorker(Map<String, Extractor<Map<String, Object>, ?>> extractorPresets) {
-        this.extractorPresets = extractorPresets;
+    JsonLoaderWorker(Map<String, Extractor<Map<String, Object>, ?>> extractors) {
+        this.extractors = extractors;
         this.appenders = new LinkedHashMap<>();
     }
 
-    protected DataFrame load(List<Map<String, Object>> parsed) {
+    DataFrame load(List<?> parsed) {
         loadColumns(parsed);
         return toDataFrame();
     }
 
-    protected void loadColumns(List<Map<String, Object>> parsed) {
+    private void loadColumns(List<?> parsed) {
 
-        // Different maps in the list can have different sets of keys...
-        // Algorithm below aligns values by row..
+        // Different maps in the list can have different sets of keys, so columns are added dynamically, and we can't
+        // use DataFrameByRowBuilder
 
         int height = parsed.size();
         int offset = 0;
@@ -50,13 +50,13 @@ class JsonLoaderWorker {
         }
     }
 
-    protected void loadRow(
+    private void loadRow(
             Map<String, Object> row,
             int height,
             int offset) {
 
         for (String column : row.keySet()) {
-            appenders.computeIfAbsent(column, label -> createdAppender(column, height, offset)).append(row);
+            appenders.computeIfAbsent(column, n -> appender(n, height, offset)).append(row);
         }
 
         // columns not in this record must be filled with nulls
@@ -69,10 +69,10 @@ class JsonLoaderWorker {
         }
     }
 
-    protected DataFrame toDataFrame() {
+    private DataFrame toDataFrame() {
         Index columnsIndex = Index.of(appenders.keySet().toArray(new String[0]));
         int w = appenders.size();
-        Series[] series = new Series[w];
+        Series<?>[] series = new Series[w];
 
         for (int i = 0; i < w; i++) {
             series[i] = appenders.get(columnsIndex.get(i)).toSeries();
@@ -81,15 +81,19 @@ class JsonLoaderWorker {
         return new ColumnDataFrame(null, columnsIndex, series);
     }
 
-    protected SeriesAppender<Map<String, Object>, ?> createdAppender(String name, int length, int offset) {
-        Extractor<Map<String, Object>, ?> extractorPreset = extractorPresets.get(name);
-        Extractor<Map<String, Object>, ?> extractor = extractorPreset != null ? extractorPreset : JsonLoader.defaultExtractor(name);
+    private SeriesAppender<Map<String, Object>, ?> appender(String name, int length, int offset) {
+        SeriesAppender<Map<String, Object>, ?> appender = new SeriesAppender<>(extractor(name), length);
 
-        SeriesAppender<Map<String, Object>, ?> appender = new SeriesAppender<>(extractor, length);
+        // since the appender may have been created in the middle of a loader run, fill the already processed
+        // positions with nulls
         for (int i = 0; i < offset; i++) {
             appender.append((Map<String, Object>) null);
         }
 
         return appender;
+    }
+
+    private Extractor<Map<String, Object>, ?> extractor(String columnName) {
+        return extractors.computeIfAbsent(columnName, n -> Extractor.$col(m -> ColConfigurator.value(m, n)));
     }
 }